Require GPTL to run performance or timing tests
LIVV now requires GPTL in order to gather performance statistics.

BATS will now parse the generated `CMakeCache.txt` build file to
determine if CISM was built with GPTL or not.

If built without GPTL, or BATS cannot determine whether CISM was built
with GPTL, BATS will issue a warning and quit. There is purposely no
over-ride of this behaviour, but, manually editing the `CMakeCache.txt`
can be used to "hack" this behavior.

This fixes ACME-Climate/LIVV#43.