Merge `-c/--compiler` option into `-p/--platform` option
Previously, BATS would use the platform and compiler options to
determine the location of the build script. For example, if the options
`-p linux -c gnu` were given, BATS would look for a `builds/linux-gnu`
directory, and a build script called `linux-gnu-cmake.bash` script
within that directory.

The current naming structure of the build directories and scripts
however isn't that rigid, and some times library and dycore
specifications are included.

Now, there is just the `-p/--platform` option, which can be given a dash
separated list of platform, compiler, dycore, and library options. For
example, if the option `-p linux-gnu-gptl` was given, it will look for a
build script named `linux-gnu-gptl-cmake.bash` in any of these
directories (in order):

    builds/linux-gnu-gptl/
    builds/linux-gnu/
    builds/linux/
    builds/

Stopping once it finds the correct build script (the first time).

This handles are current directory and script combinations within CISM.